Added Foreign Keys and appropriate annotations and also changed the primary key's name

+++ src/main/java/Master/Country_Master/Dao/Country_Dao.java
@@ -5,6 +5,6 @@ import Master.Country_Master.entity.Country_Model;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface Country_Dao extends JpaRepository<Country_Model, Long>{
+public interface Country_Dao extends JpaRepository<Country_Model, Integer>{
 
 }
